ref: Moved icon components to adequate directory
