--- src/components/MoreDetailsIcon.tsx
@@ -1,7 +1,0 @@
-export default function MoreDetailsIcon() {
-    return (
-        <svg width="16" height="16" viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg">
-            <path fill="currentColor" d="M4.40933 3.29103L9.11933 8.00103L4.40933 12.711C4.3615 12.7559 4.32338 12.8101 4.29732 12.8703C4.27126 12.9305 4.25781 12.9954 4.25781 13.061C4.25781 13.1266 4.27126 13.1915 4.29732 13.2517C4.32338 13.3119 4.3615 13.3661 4.40933 13.411L5.11933 14.121C5.16422 14.1689 5.21844 14.207 5.27864 14.233C5.33883 14.2591 5.40374 14.2726 5.46933 14.2726C5.53493 14.2726 5.59983 14.2591 5.66003 14.233C5.72022 14.207 5.77444 14.1689 5.81933 14.121L11.5893 8.35103C11.6372 8.30614 11.6753 8.25193 11.7013 8.19173C11.7274 8.13153 11.7409 8.06663 11.7409 8.00103C11.7409 7.93544 11.7274 7.87054 11.7013 7.81034C11.6753 7.75014 11.6372 7.69592 11.5893 7.65103L5.81933 1.88103C5.77444 1.8332 5.72022 1.79508 5.66003 1.76902C5.59983 1.74296 5.53493 1.72952 5.46933 1.72952C5.40373 1.72952 5.33883 1.74296 5.27864 1.76902C5.21844 1.79508 5.16422 1.8332 5.11933 1.88103L4.40933 2.59103C4.3615 2.63592 4.32338 2.69014 4.29732 2.75034C4.27126 2.81054 4.25781 2.87544 4.25781 2.94103C4.25781 3.00663 4.27126 3.07153 4.29732 3.13173C4.32338 3.19193 4.3615 3.24614 4.40933 3.29103Z"/>
-        </svg>
-    )
-}
